break out jobs management into its own crate
