+++ Cargo.toml
@@ -40,11 +40,7 @@ members = [
 	"node/core/proposer",
 	"node/core/provisioner",
 	"node/core/runtime-api",
-	"node/jaeger",
-	"node/network/approval-distribution",
-	"node/network/availability-distribution",
-	"node/network/availability-recovery",
-	"node/network/bitfield-distribution",
+	"node/jobs",
 	"node/network/bridge",
 	"node/network/collator-protocol",
 	"node/network/pov-distribution",
