+++ node/core/backing/Cargo.toml
@@ -9,6 +9,7 @@ futures = "0.3.12"
 sp-keystore = { git = "https://github.com/paritytech/substrate", branch = "master" }
 polkadot-primitives = { path = "../../../primitives" }
 pnu-primitives = { path = "../../primitives" }
+pnu-jobs = { path = "../../jobs" }
 polkadot-subsystem = { package = "pnu-subsystem", path = "../../subsystem" }
 pnu-subsystem-util = { path = "../../subsystem-util" }
 erasure-coding = { package = "polkadot-erasure-coding", path = "../../../erasure-coding" }
