+++ node/core/candidate-selection/Cargo.toml
@@ -13,6 +13,7 @@ thiserror = "1.0.23"
 sp-keystore = { git = "https://github.com/paritytech/substrate", branch = "master" }
 
 polkadot-primitives = { path = "../../../primitives" }
+pnu-jobs = { path = "../../jobs" }
 pnu-subsystem = { path = "../../subsystem" }
 pnu-subsystem-util = { path = "../../subsystem-util" }
