+++ node/subsystem-util/Cargo.toml
@@ -12,28 +12,16 @@ futures-timer = "3.0.2"
 parity-scale-codec = { version = "1.3.6", default-features = false, features = ["derive"] }
 parking_lot = { version = "0.11.1", optional = true }
 pin-project = "1.0.4"
-streamunordered = "0.5.1"
 thiserror = "1.0.23"
 tracing = "0.1.22"
-tracing-futures = "0.2.4"
 
+pnu-metered-channel = { path = "../util/metered-channel"}
 pnu-primitives = { path = "../primitives" }
 pnu-subsystem = { path = "../subsystem" }
-pnu-jaeger = { path = "../jaeger" }
 polkadot-primitives = { path = "../../primitives" }
-pnu-metered-channel = { path = "../util/metered-channel"}
 
 sc-network = { git = "https://github.com/paritytech/substrate", branch = "master" }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-application-crypto = { git = "https://github.com/paritytech/substrate", branch = "master" }
+sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-keystore = { git = "https://github.com/paritytech/substrate", branch = "master" }
 substrate-prometheus-endpoint = { git = "https://github.com/paritytech/substrate", branch = "master" }
-
-[dev-dependencies]
-assert_matches = "1.4.0"
-async-trait = "0.1.42"
-env_logger = "0.8.2"
-futures = { version = "0.3.12", features = ["thread-pool"] }
-log = "0.4.13"
-parking_lot = "0.11.1"
-pnu-subsystem-test-helpers = { path = "../subsystem-test-helpers" }
